+++ python/dgl/distributed/dist_graph.py
@@ -14,12 +14,14 @@ from .._ffi.ndarray import empty_shared_mem
 from ..frame import infer_scheme
 from .partition import load_partition
 from .graph_partition_book import PartitionPolicy, get_shared_mem_partition_book
+from .graph_partition_book import NODE_PART_POLICY, NODE_PART_POLICY
 from .. import utils
 from .shared_mem_utils import _to_shared_mem, _get_ndata_path, _get_edata_path, DTYPE_DICT
 from . import rpc
 from .rpc_client import connect_to_server
 from .server_state import ServerState
 from .rpc_server import start_server
+from .dist_tensor import DistTesnor
 from ..transform import as_heterograph
 
 def _get_graph_path(graph_name):
@@ -42,15 +44,6 @@ FIELD_DICT = {'inner_node': F.int64,
               'inner_edge': F.int64,
               NID: F.int64,
               EID: F.int64}
-NODE_PART_POLICY = 'node'
-EDGE_PART_POLICY = 'edge'
-
-def _get_data_name(name, part_policy):
-    ''' This is to get the name of data in the kvstore.
-
-    KVStore doesn't understand node data or edge data. We'll use a prefix to distinguish them.
-    '''
-    return part_policy + ':' + name
 
 def _is_ndata_name(name):
     ''' Is this node data in the kvstore '''
@@ -103,95 +96,6 @@ def _get_graph_from_shared_mem(graph_name):
     g.ndata[NID] = _get_shared_mem_ndata(g, graph_name, NID)
     g.edata[EID] = _get_shared_mem_edata(g, graph_name, EID)
     return g
-
-class DistTensor:
-    ''' Distributed tensor.
-
-    This is a wrapper to access a tensor stored in the distributed KVStore.
-    This wrapper provides an interface similar to the local tensor.
-
-    If a user tries to create a new tensor with a name that exists in the KVStore,
-    the creation will fail.
-
-    Parameters
-    ----------
-    g : DistGraph
-        The distributed graph object.
-    shape : tuple
-        The shape of the tensor
-    dtype : dtype
-        The dtype of the tensor
-    name : string
-        The name of the tensor.
-    part_policy : PartitionPolicy
-        The partition policy of the tensor
-    init_func : callable
-        The function to initialize data in the tensor.
-    create_new : bool
-        Whether or not to create a new tensor in the KVStore.
-    '''
-    def __init__(self, g, shape, dtype, name, part_policy=None, init_func=None,
-                 create_new=True):
-        self.kvstore = g._client
-        self._shape = shape
-        self._dtype = dtype
-
-        if part_policy is None:
-            assert shape[0] != g.number_of_nodes() or shape[0] != g.number_of_edges(), \
-                    'Cannot determine the partition policy. Please provide it.'
-            if shape[0] == g.number_of_nodes():
-                part_policy = PartitionPolicy(NODE_PART_POLICY, g.get_partition_book())
-            elif shape[0] == g.number_of_edges():
-                part_policy = PartitionPolicy(EDGE_PART_POLICY, g.get_partition_book())
-            else:
-                raise DGLError('Cannot determine the partition policy. Please provide it.')
-
-        self._part_policy = part_policy
-
-        if init_func is None:
-            init_func = _default_init_data
-        self._name = _get_data_name(name, part_policy.policy_str)
-        if create_new:
-            g._client.init_data(self._name, shape, dtype, part_policy, init_func)
-        else:
-            dtype1, shape1, _ = g._client.get_data_meta(self._name)
-            assert dtype == dtype1
-            assert shape == shape1
-
-    def __getitem__(self, idx):
-        idx = utils.toindex(idx)
-        idx = idx.tousertensor()
-        return self.kvstore.pull(name=self._name, id_tensor=idx)
-
-    def __setitem__(self, idx, val):
-        idx = utils.toindex(idx)
-        idx = idx.tousertensor()
-        # TODO(zhengda) how do we want to support broadcast (e.g., G.ndata['h'][idx] = 1).
-        self.kvstore.push(name=self._name, id_tensor=idx, data_tensor=val)
-
-    def __len__(self):
-        return self._shape[0]
-
-    @property
-    def part_policy(self):
-        ''' Return the partition policy '''
-        return self._part_policy
-
-    @property
-    def shape(self):
-        ''' Return the shape of the distributed tensor. '''
-        return self._shape
-
-    @property
-    def dtype(self):
-        ''' Return the data type of the distributed tensor. '''
-        return self._dtype
-
-    @property
-    def name(self):
-        ''' Return the name of the distributed tensor '''
-        return self._name
-
 
 class NodeDataView(MutableMapping):
     """The data view class when dist_graph.ndata[...].data is called.
